requirements.txt
asgiref==3.7.2
attrs==23.1.0
bleach==6.1.0
Django==4.2.8
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3
django-jazzmin==2.6.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.9.1
drf-yasg==1.21.7
environ==1.0
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
packaging==23.22.9.9
Pillow==10.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
webencodings==0.5.1
